LayoutTests/fast/text-autosizing/ios/percent-adjust-length-line-height-expected.txt

Tests length line-height adjustment for percentage -webkit-text-size-adjust values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedLineHeight is '24px'
PASS successfullyParsed is true

TEST COMPLETE
